+++ BenchmarkSample/ui/src/main/java/com/example/benchmark/ui/MainActivity.kt
@@ -1,17 +1,16 @@
 package com.example.benchmark.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import kotlin.random.Random.Default.nextLong
 
 class MainActivity : AppCompatActivity() {
+
+    // It's used in benchmarks
+    @Suppress("MemberVisibilityCanBePrivate")
     val adapter = SampleAdapter()
+
     lateinit var recyclerView: RecyclerView
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,35 +23,7 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
-class Holder(root: View) : RecyclerView.ViewHolder(root) {
-    private val textView: TextView = root.findViewById(R.id.label)
-
-    fun bind(string: String) {
-        textView.text = string
-    }
-
-    companion object {
-        fun create(parent: ViewGroup): Holder {
-            val inflater = LayoutInflater.from(parent.context)
-            return Holder(inflater.inflate(R.layout.item_view, parent, false))
-        }
-    }
-}
-
-class SampleAdapter : ListAdapter<String, Holder>(StringDiffCallback) {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
-        return Holder.create(parent)
-    }
-
-    override fun onBindViewHolder(holder: Holder, position: Int) {
-        holder.bind(getItem(position))
-    }
-}
-
-private object StringDiffCallback : DiffUtil.ItemCallback<String>() {
-    override fun areItemsTheSame(oldItem: String, newItem: String): Boolean =
-        oldItem == newItem
-
-    override fun areContentsTheSame(oldItem: String, newItem: String): Boolean =
-        oldItem == newItem
+fun doSomeWork() {
+    // Pretend this method does some work
+    Thread.sleep(nextLong(1_000))
 }
